protecting notes with auth middleware

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,7 +19,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
     }
@@ -22,6 +30,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, 'Note not found');
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
+    }
+
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -40,11 +53,19 @@ export const createNote: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const { title, text } = req.body;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, 'Note must have a title');
     }
-    const newNote = await NoteModel.create({ title, text });
+    const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
+      title,
+      text,
+    });
     res.status(201).json(newNote);
   } catch (error) {
     next(error);
@@ -68,8 +89,11 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
     }
@@ -80,6 +104,11 @@ export const updateNote: RequestHandler<
     if (!note) {
       throw createHttpError(404, 'Note not found');
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
+    }
+
     note.title = newTitle;
     note.text = newText;
 
@@ -92,13 +121,21 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note ID');
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
 
     await note.deleteOne();
